# internal/docs/swagger.yaml
definitions:
...
  api.AssetResponse:
    properties:
      classification_category:
        type: string
      classification_status:
        type: string
      created_at:
        type: string
      derived_from_asset_id:
        type: string
      duplicate_of:
        allOf:
        - $ref: '#/definitions/api.DuplicateOfResponse'
        description: |-
          DuplicateOf is set only on the POST /assets response, when the
          workspace's duplicate_detection_mode is "warn" and a content-hash match
          was found elsewhere in the workspace. Omitted entirely (not null) when
          there is no match, and never populated by GET/list responses.
      folder_id:
        type: string
      height:
        type: integer
      id:
        type: string
      metadata:
        type: string
      mime_type:
        type: string
      original_filename:
        type: string
      project_id:
        type: string
      shared_variants:
        items:
          $ref: '#/definitions/api.SharedVariantResponse'
        type: array
      size:
        type: integer
      storage_key:
        description: |-
          StorageKey is returned to authenticated workspace clients so the desktop
          shell can map an asset to the same shared-storage root used by the server.
          Public share responses deliberately leave it empty.
        type: string
      tags:
        items:
          type: string
        type: array
      thumbnail_content_type:
        type: string
      thumbnail_key:
        type: string
      updated_at:
        type: string
      variant_count:
        type: integer
      variants_rebuilding:
        type: boolean
      version_count:
        type: integer
      width:
        type: integer
      workspace_id:
        type: string
    required:
    - classification_status
    - created_at
    - id
    - mime_type
    - original_filename
    - size
    - tags
    - updated_at
    - variant_count
    - variants_rebuilding
    - version_count
    - workspace_id
    type: object
...
  api.DuplicateOfResponse:
    properties:
      asset_id:
        type: string
      created_at:
        type: string
      is_deleted_version:
        type: boolean
      original_filename:
        type: string
      project_id:
        type: string
      storage_available:
        type: boolean
      thumbnail_url:
        type: string
      version_id:
        type: string
    required:
    - asset_id
    - created_at
    - is_deleted_version
    - original_filename
    - storage_available
    - version_id
    type: object
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    required:
    - error
    type: object
...
  api.SharedVariantResponse:
    properties:
      id:
        type: string
      mime_type:
        type: string
      size:
        type: integer
      thumbnail_content_type:
        type: string
      thumbnail_url:
        type: string
      title:
        type: string
      type:
        type: string
    required:
    - id
    - mime_type
    - thumbnail_content_type
    - title
    - type
    type: object
...
  api.VersionCreatedByResponse:
    properties:
      id:
        type: string
      name:
        type: string
    required:
    - id
    - name
    type: object
  api.VersionResponse:
    properties:
      comment:
        type: string
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/api.VersionCreatedByResponse'
      duration_sec:
        type: number
      height:
        type: integer
      id:
        type: string
      is_current:
        type: boolean
      mime_type:
        type: string
      size:
        type: integer
      thumbnail_url:
        type: string
      variant_count:
        type: integer
      version_num:
        type: integer
      width:
        type: integer
    required:
    - created_at
    - id
    - is_current
    - mime_type
    - size
    - variant_count
    - version_num
    type: object
  api.VersionWithAssetResponse:
    properties:
      asset:
        $ref: '#/definitions/api.AssetResponse'
      version:
        $ref: '#/definitions/api.VersionResponse'
    required:
    - asset
    - version
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/assets/{id}/versions:
    get:
      description: 'Returns all versions of an asset in reverse chronological order.
        Each version includes its <code>version_num</code>, MIME type, dimensions,
        file size, optional comment, and the number of variants derived from it (<code>variant_count</code>).
        The current version has <code>is_current: true</code>.'
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.VersionResponse'
            type: array
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Asset not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List asset versions
      tags:
      - Versions
    post:
      consumes:
      - multipart/form-data
      description: 'Uploads a new file as the next version of an existing asset. The
        new version is immediately promoted to current. The previous current version
        is retained in history.<br><br> After upload, the server: <ul> <li>Enqueues
        a thumbnail generation job for the new version.</li> <li>Enqueues a variant
        rebuild job that recreates all variants from the previous current version.</li>
        <li>Clears the asset''s cached thumbnail (pending the new job).</li> </ul>
        If the uploaded file is byte-for-byte identical to the current version (same
        content hash), the request is rejected with 409 to prevent no-op versions.<br><br>
        The response body contains both the new <code>version</code> object and the
        updated <code>asset</code> object.'
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      - description: New version file
        in: formData
        name: file
        required: true
        type: file
      - description: Optional change note (max 500 characters)
        in: formData
        name: comment
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.VersionWithAssetResponse'
        "400":
          description: Missing file field
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Asset not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: File is identical to the current version
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Comment too long
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload a new version
      tags:
      - Versions
  /api/v1/assets/{id}/versions/{vid}:
    delete:
      description: Soft-deletes an asset version. The current version cannot be deleted
        — restore a different version first, then delete the old one. A version also
        cannot be deleted while it is in use as a project cover or workspace icon
        (returns 409).<br><br> Soft deletion marks the row as deleted; the storage
        file is not immediately removed (physical cleanup is handled by a background
        retention job).
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      - description: Version ID
        in: path
        name: vid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Asset or version not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Version is in use as a cover/icon
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Cannot delete the current version
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a version
      tags:
      - Versions
  /api/v1/assets/{id}/versions/{vid}/file:
    get:
      description: Streams the raw file for the specified version. The response MIME
        type and <code>Content-Disposition</code> are set from the stored version
        metadata.
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      - description: Version ID
        in: path
        name: vid
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Asset, version, or file not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get version file
      tags:
      - Versions
  /api/v1/assets/{id}/versions/{vid}/restore:
    post:
      description: Promotes a non-current version to be the current version of the
        asset. The asset's thumbnail is also updated to match the restored version.
        An audit event records the version numbers rolled back from and to.<br><br>
        Returns 409 if the requested version is already current. Returns 422 if the
        version has been soft-deleted. The response contains the updated <code>version</code>
        and <code>asset</code>.
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      - description: Version ID to restore
        in: path
        name: vid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.VersionWithAssetResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Asset or version not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Version is already current
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Cannot restore a deleted version
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Restore a version
      tags:
      - Versions
  /api/v1/assets/{id}/versions/{vid}/thumb:
    get:
      description: Returns the JPEG thumbnail for the given version. Returns 404 if
        the thumbnail has not yet been generated (the job runs asynchronously after
        upload — poll until the <code>thumbnail_url</code> field in the version response
        is non-null).
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      - description: Version ID
        in: path
        name: vid
        required: true
        type: string
      produces:
      - image/jpeg
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Asset, version, or thumbnail not ready
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get version thumbnail
      tags:
      - Versions
...
swagger: "2.0"